+++ AoE2ScenarioParser/datasets/techs.py
@@ -1,5 +1,11 @@
 from __future__ import annotations
 from enum import Enum
+from typing import Union
+
+from AoE2ScenarioParser.datasets.buildings import BuildingInfo
+
+from AoE2ScenarioParser.datasets.trigger_lists import Age
+from AoE2ScenarioParser.helper.list_functions import listify
 
 
 class TechInfo(Enum):
@@ -15,6 +21,11 @@ class TechInfo(Enum):
     >>> TechInfo.from_icon_id()
     >>> TechInfo.unique_techs()
     >>> TechInfo.unique_unit_upgrades()
+    >>> TechInfo.town_center_techs()
+    >>> TechInfo.blacksmith_techs()
+    >>> TechInfo.monastery_techs()
+    >>> TechInfo.university_techs()
+    >>> TechInfo.eco_techs()
 
     **Examples**
 
@@ -24,8 +35,9 @@ class TechInfo(Enum):
     >>> TechInfo.LOOM.ICON_ID
     >>> 6
     """
+
     @property
-    def ID(self):
+    def ID(self) -> int:
         """
         Returns:
             The ID of the specified tech
@@ -52,7 +64,7 @@ class TechInfo(Enum):
         raise KeyError(f"A technology with ID '{tech_id}' was not found in the dataset")
 
     @property
-    def ICON_ID(self):
+    def ICON_ID(self) -> int:
         """
         Returns:
             The icon ID of the specified tech
@@ -60,7 +72,7 @@ class TechInfo(Enum):
         return self.value[1]
 
     @classmethod
-    def from_icon_id(cls, tech_icon_id):
+    def from_icon_id(cls, tech_icon_id) -> TechInfo:
         """
         Get the TechInfo object from its icon ID
 
@@ -186,8 +198,8 @@ class TechInfo(Enum):
 
     @staticmethod
     def unique_unit_upgrades(
-        exclude_castle_techs: bool = False,
-        exclude_non_castle_techs: bool = False
+            exclude_castle_techs: bool = False,
+            exclude_non_castle_techs: bool = False
     ) -> list[TechInfo]:
         """
         Args:
@@ -257,6 +269,225 @@ class TechInfo(Enum):
             techs_to_return.extend(unique_techs["castle"])
         if not exclude_non_castle_techs:
             techs_to_return.extend(unique_techs["non_castle"])
+
+        return techs_to_return
+
+    @staticmethod
+    def town_center_techs(ages: Union[int, list[int]] = None):
+        """
+        Args:
+            ages: a list of age IDs (IDs are located in the Age IntEnum dataset). If specified, only techs from these
+                ages are returned. If unspecified, all ages' techs are returned
+
+        Returns:
+            A list of TechInfo objects which are the blacksmith upgrade techs for the given age
+        """
+        ages = list(Age) if ages is None else listify(ages)
+
+        upgrades = {
+            Age.DARK_AGE: [
+                TechInfo.LOOM,
+                TechInfo.FEUDAL_AGE
+            ],
+            Age.FEUDAL_AGE: [
+                TechInfo.WHEELBARROW,
+                TechInfo.TOWN_WATCH,
+                TechInfo.CASTLE_AGE
+            ],
+            Age.CASTLE_AGE: [
+                TechInfo.HAND_CART,
+                TechInfo.TOWN_PATROL,
+                TechInfo.IMPERIAL_AGE
+            ],
+            Age.IMPERIAL_AGE: [],
+        }
+
+        techs_to_return = []
+        for age in ages:
+            techs_to_return.extend(upgrades[age])
+
+        return techs_to_return
+
+    @staticmethod
+    def blacksmith_techs(ages: Union[int, list[int]] = None) -> list[TechInfo]:
+        """
+        Args:
+            ages: a list of age IDs (IDs are located in the Age IntEnum dataset). If specified, only techs from these
+                ages are returned. If unspecified, all ages' techs are returned
+
+        Returns:
+            A list of TechInfo objects which are the blacksmith upgrade techs for the given age
+        """
+        ages = list(Age) if ages is None else listify(ages)
+
+        upgrades = {
+            Age.DARK_AGE: [],
+            Age.FEUDAL_AGE: [
+                TechInfo.FORGING,
+                TechInfo.SCALE_MAIL_ARMOR,
+                TechInfo.SCALE_BARDING_ARMOR,
+                TechInfo.FLETCHING,
+                TechInfo.PADDED_ARCHER_ARMOR
+            ],
+            Age.CASTLE_AGE: [
+                TechInfo.IRON_CASTING,
+                TechInfo.CHAIN_MAIL_ARMOR,
+                TechInfo.CHAIN_BARDING_ARMOR,
+                TechInfo.BODKIN_ARROW,
+                TechInfo.LEATHER_ARCHER_ARMOR
+            ],
+            Age.IMPERIAL_AGE: [
+                TechInfo.BLAST_FURNACE,
+                TechInfo.PLATE_MAIL_ARMOR,
+                TechInfo.PLATE_BARDING_ARMOR,
+                TechInfo.BRACER,
+                TechInfo.RING_ARCHER_ARMOR
+            ],
+        }
+
+        techs_to_return = []
+        for age in ages:
+            techs_to_return.extend(upgrades[age])
+
+        return techs_to_return
+
+    @staticmethod
+    def monastery_techs(ages: Union[int, list[int]] = None) -> list[TechInfo]:
+        """
+        Args:
+            ages: The age ID (IDs are located in the Age IntEnum dataset). If specified, only techs from these
+                ages are returned. If unspecified, all ages' techs are returned
+
+        Returns:
+            A list of TechInfo objects which are the monastery upgrade techs for the given age
+        """
+        ages = list(Age) if ages is None else listify(ages)
+
+        upgrades = {
+            Age.DARK_AGE: [],
+            Age.FEUDAL_AGE: [],
+            Age.CASTLE_AGE: [
+                TechInfo.REDEMPTION,
+                TechInfo.ATONEMENT,
+                TechInfo.HERBAL_MEDICINE,
+                TechInfo.HERESY,
+                TechInfo.SANCTITY,
+                TechInfo.FERVOR,
+            ],
+            Age.IMPERIAL_AGE: [
+                TechInfo.FAITH,
+                TechInfo.ILLUMINATION,
+                TechInfo.BLOCK_PRINTING,
+                TechInfo.THEOCRACY
+            ],
+        }
+
+        techs_to_return = []
+        for age in ages:
+            techs_to_return.extend(upgrades[age])
+
+        return techs_to_return
+
+    @staticmethod
+    def university_techs(ages: Union[int, list[int]] = None) -> list[TechInfo]:
+        """
+        Args:
+            ages: The age ID (IDs are located in the Age IntEnum dataset). If specified, only techs from these
+                ages are returned. If unspecified, all ages' techs are returned
+
+        Returns:
+            A list of TechInfo objects which are the university upgrade techs for the given age
+        """
+        ages = list(Age) if ages is None else listify(ages)
+
+        upgrades = {
+            Age.DARK_AGE: [],
+            Age.FEUDAL_AGE: [],
+            Age.CASTLE_AGE: [
+                TechInfo.MASONRY,
+                TechInfo.FORTIFIED_WALL,
+                TechInfo.BALLISTICS,
+                TechInfo.GUARD_TOWER,
+                TechInfo.HEATED_SHOT,
+                TechInfo.MURDER_HOLES,
+                TechInfo.TREADMILL_CRANE,
+            ],
+            Age.IMPERIAL_AGE: [
+                TechInfo.ARCHITECTURE,
+                TechInfo.CHEMISTRY,
+                TechInfo.SIEGE_ENGINEERS,
+                TechInfo.KEEP,
+                TechInfo.ARROWSLITS,
+                TechInfo.BOMBARD_TOWER
+            ],
+        }
+
+        techs_to_return = []
+        for age in ages:
+            techs_to_return.extend(upgrades[age])
+
+        return techs_to_return
+
+    @staticmethod
+    def eco_techs(ages: Union[int, list[int]] = None, buildings: Union[int, list[int]] = None) -> list[TechInfo]:
+        """
+        Args:
+            ages: The age ID (IDs are located in the Age IntEnum dataset). If specified, only techs from these
+                ages are returned. If unspecified, all ages' techs are returned
+            buildings: The building ID for which the upgrades are needed. If unspecified, eco upgrades from all
+                economic buildings (Mill, Mining Camp, Lumber Camp, Town Center and Market) are returned
+
+        Returns:
+            A list of TechInfo objects which are the university upgrade techs for the given age
+        """
+        ages = list(Age) if ages is None else listify(ages)
+
+        if buildings is None:
+            buildings = [
+                BuildingInfo.MILL.ID,
+                BuildingInfo.MINING_CAMP.ID,
+                BuildingInfo.LUMBER_CAMP.ID,
+                BuildingInfo.TOWN_CENTER.ID,
+                BuildingInfo.MARKET.ID
+            ]
+        else:
+            buildings = listify(buildings)
+
+        upgrades = {
+            Age.DARK_AGE: {
+                BuildingInfo.MILL.ID: [],
+                BuildingInfo.MINING_CAMP.ID: [],
+                BuildingInfo.LUMBER_CAMP.ID: [],
+                BuildingInfo.TOWN_CENTER.ID: [],
+                BuildingInfo.MARKET.ID: [],
+            },
+            Age.FEUDAL_AGE: {
+                BuildingInfo.MILL.ID: [TechInfo.HORSE_COLLAR],
+                BuildingInfo.MINING_CAMP.ID: [TechInfo.GOLD_MINING, TechInfo.STONE_MINING],
+                BuildingInfo.LUMBER_CAMP.ID: [TechInfo.DOUBLE_BIT_AXE],
+                BuildingInfo.TOWN_CENTER.ID: [TechInfo.WHEELBARROW],
+                BuildingInfo.MARKET.ID: [],
+            },
+            Age.CASTLE_AGE: {
+                BuildingInfo.MILL.ID: [TechInfo.HEAVY_PLOW],
+                BuildingInfo.MINING_CAMP.ID: [TechInfo.GOLD_SHAFT_MINING, TechInfo.STONE_SHAFT_MINING],
+                BuildingInfo.LUMBER_CAMP.ID: [TechInfo.BOW_SAW],
+                BuildingInfo.TOWN_CENTER.ID: [TechInfo.HAND_CART],
+                BuildingInfo.MARKET.ID: [TechInfo.COINAGE, TechInfo.CARAVAN]
+            },
+            Age.IMPERIAL_AGE: {
+                BuildingInfo.MILL.ID: [TechInfo.CROP_ROTATION],
+                BuildingInfo.MINING_CAMP.ID: [],
+                BuildingInfo.LUMBER_CAMP.ID: [TechInfo.TWO_MAN_SAW],
+                BuildingInfo.TOWN_CENTER.ID: [],
+                BuildingInfo.MARKET.ID: [TechInfo.BANKING, TechInfo.GUILDS]
+            },
+        }
+
+        techs_to_return = []
+        for age in ages:
+            for building in buildings:
+                techs_to_return.extend(upgrades[age][building])
 
         return techs_to_return