+++ AoE2ScenarioParser/scenarios/aoe2_de_scenario.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from AoE2ScenarioParser.objects.managers.de.map_manager_de import MapManagerDE
 from AoE2ScenarioParser.objects.managers.de.trigger_manager_de import TriggerManagerDE
 from AoE2ScenarioParser.objects.managers.de.unit_manager_de import UnitManagerDE
+from AoE2ScenarioParser.objects.managers.de.xs_manager_de import XsManagerDE
 from AoE2ScenarioParser.scenarios.aoe2_scenario import AoE2Scenario
 
 
@@ -18,6 +19,10 @@ class AoE2DEScenario(AoE2Scenario):
     @property
     def map_manager(self) -> MapManagerDE:
         return self._object_manager.managers['Map']
+
+    @property
+    def xs_manager(self) -> XsManagerDE:
+        return self._object_manager.managers['Xs']
 
     @classmethod
     def from_file(cls, filename, game_version="DE") -> AoE2DEScenario: